requirements.txt
asgiref==3.8.1
boto3==1.34.140
botocore==1.34.140
certifi==2024.7.4
decouple==0.0.7
dj-database-url==2.2.0
Django==5.0.6
django-cors-headers==4.4.0
djangorestframework==3.15.2
jmespath==1.0.1
psycopg2-binary==2.9.9
PyJWT==2.8.0
python-dateutil==2.9.0.post0
python-decouple==3.8
python-http-client==3.3.7
s3transfer==0.10.2
sendgrid==6.11.0
six==1.16.0
sqlparse==0.5.0
starkbank-ecdsa==2.2.0
typing_extensions==4.12.2
urllib3==2.2.2
twilio==9.3.3
python-dotenv==1.0.1